feat: move drawer to dedicated component
